BUGFIX in squid backend

+++ imswitch/imcontrol/model/managers/positioners/SQUIDStageManager.py
@@ -18,11 +18,11 @@ class SQUIDStageManager(PositionerManager):
 
     def move(self, value, axis):
         if axis == 'X':
-            self._rs232manager._squid.move_x_usteps(value*self.PHYS_FACTOR)
+            self._rs232manager._squid.move_x_usteps(int(value*self.PHYS_FACTOR))
         elif axis == 'Y':
-            self._rs232manager._squid.move_y_usteps(value*self.PHYS_FACTOR)
+            self._rs232manager._squid.move_y_usteps(int(value*self.PHYS_FACTOR))
         elif axis == 'Z':
-            self._rs232manager._squid.move_z_usteps(value*self.PHYS_FACTOR)
+            self._rs232manager._squid.move_z_usteps(int(value*self.PHYS_FACTOR))
         else:
             print('Wrong axis, has to be "X" "Y" or "Z".')
             return
